@vtjnash 2021-03-16
Initial test run of Nanosoldier (on a somewhat noisy test machine) with this PR seems to be showing substantial improvement in load and net runtime without affecting benchmark time. Will be 8 hours until the updated result with ALL finishes (don't tag a release before then or it'll mess up the comparison!): https://github.com/vtjnash/julia/commit/d7da2a4a327b47f4969b40a9feadef3b2b825eda