@vtjnash 2021-03-17
ALL report finished, and shows this might also be up to 25% faster end-to-end, with no regressions on any particular benchmark on this PR branch, when run against BenchmarkTools v0.6.0 (current master)